+++ tic-tac-toe/components/Square.js
@@ -1,46 +1,22 @@
-import {
-  dom as d,
-  createStore,
-  register
-} from '../../todomvc/lib/state-manager';
+import { dom as d } from '../../todomvc/lib/state-manager';
 
 // Square
 // <button class="square">
-//   {/* TODO */}
+//   value
 // </button>
-function Square() {
-  // Returns a store with a dispatch function to
-  // trigger state changes by dispatching actions.
-  const store = createStore({
-    value: null
-  });
-
-  function view(state) {
-    return d(
-      'button',
-      {
-        class: 'square',
-        on: {
-          click: () => {
-            store.dispatch(
-              {
-                state: { value: 'X' }
-              },
-              true
-            );
-          }
+function Square({ value, onClick }) {
+  return d(
+    'button',
+    {
+      class: 'square',
+      on: {
+        click: () => {
+          onClick();
         }
-      },
-      state.value
-    );
-  }
-
-  const component = register({
-    view,
-    store
-  });
-
-  return component;
+      }
+    },
+    value
+  );
 }
 
 export default Square;
